test: decouple lifespan-aware HTTP support
