+++ pyproject.toml
@@ -22,6 +22,8 @@ dependencies = [
 
 [project.optional-dependencies]
 dev = [
+    "asgi-lifespan>=2.1.0",
+    "httpx2>=2.0.0",
     "mypy>=2.2.0",
     "ruff>=0.15.21",
 ]
