+++ tests/support/__init__.py
@@ -1,0 +1,1 @@
+"""Narrow, reusable support for Launchplane test contracts."""
