+++ tests/test_http_app_records.py
@@ -94,11 +94,9 @@ from tests.http_app_test_support import (
     _write_recent_operations_records,
     _write_secret_status_records,
 )
-from tests.test_service import (
-    _identity,
-    _sqlite_database_url,
-    _StubVerifier,
-)
+from tests.support.raw_asgi import request as raw_asgi_request
+from tests.support.auth import _identity, _StubVerifier
+from tests.support.stores import _sqlite_database_url
 
 
 class FastApiDeploymentPromotionReadTests(unittest.IsolatedAsyncioTestCase):
@@ -3082,7 +3080,7 @@ class FastApiDeploymentEvidenceTests(unittest.IsolatedAsyncioTestCase):
             record_store_factory=lambda: _MissingProductReadStore(),
         )
 
-        response = await _asgi_request(
+        response = await raw_asgi_request(
             app,
             "POST",
             "/v1/evidence/deployments",
@@ -3133,7 +3131,7 @@ class FastApiDeploymentEvidenceTests(unittest.IsolatedAsyncioTestCase):
             record_store_factory=lambda: _MissingProductReadStore(),
         )
 
-        response = await _asgi_request(
+        response = await raw_asgi_request(
             app,
             "POST",
             "/v1/evidence/deployments",
@@ -3161,7 +3159,7 @@ class FastApiDeploymentEvidenceTests(unittest.IsolatedAsyncioTestCase):
             record_store_factory=lambda: _MissingProductReadStore(),
         )
 
-        response = await _asgi_request(
+        response = await raw_asgi_request(
             app,
             "POST",
             "/v1/evidence/deployments",
@@ -3189,7 +3187,7 @@ class FastApiDeploymentEvidenceTests(unittest.IsolatedAsyncioTestCase):
             record_store_factory=lambda: _MissingProductReadStore(),
         )
 
-        response = await _asgi_request(
+        response = await raw_asgi_request(
             app,
             "POST",
             "/v1/evidence/deployments",
@@ -3217,7 +3215,7 @@ class FastApiDeploymentEvidenceTests(unittest.IsolatedAsyncioTestCase):
             record_store_factory=lambda: _MissingProductReadStore(),
         )
 
-        response = await _asgi_request(
+        response = await raw_asgi_request(
             app,
             "POST",
             "/v1/evidence/deployments",
@@ -3246,7 +3244,7 @@ class FastApiDeploymentEvidenceTests(unittest.IsolatedAsyncioTestCase):
             record_store_factory=lambda: _MissingProductReadStore(),
         )
 
-        response = await _asgi_request(
+        response = await raw_asgi_request(
             app,
             "POST",
             "/v1/evidence/deployments",
@@ -3274,7 +3272,7 @@ class FastApiDeploymentEvidenceTests(unittest.IsolatedAsyncioTestCase):
             record_store_factory=lambda: _MissingProductReadStore(),
         )
 
-        response = await _asgi_request(
+        response = await raw_asgi_request(
             app,
             "POST",
             "/v1/evidence/deployments",
@@ -3302,7 +3300,7 @@ class FastApiDeploymentEvidenceTests(unittest.IsolatedAsyncioTestCase):
             record_store_factory=lambda: _MissingProductReadStore(),
         )
 
-        response = await _asgi_request(
+        response = await raw_asgi_request(
             app,
             "POST",
             "/v1/evidence/deployments",
@@ -3332,7 +3330,7 @@ class FastApiDeploymentEvidenceTests(unittest.IsolatedAsyncioTestCase):
             record_store_factory=lambda: _MissingProductReadStore(),
         )
 
-        response = await _asgi_request(
+        response = await raw_asgi_request(
             app,
             "POST",
             "/v1/evidence/deployments",
